homework_review/requirements.txt
python-docx>=1.0.0
openpyxl>=3.0.0
requests>=2.28.0
python-dotenv>=0.19.0
